% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANZSCO.R
\docType{package}
\name{ANZSCO-package}
\alias{ANZSCO-package}
\alias{ANZSCO}
\title{ANZSCO definitions}
\description{
Provides fully defined ANZSCO occupations in long form.
}
\details{
ANZSCO-package
A data package for ANZSCO.

Use \code{data(package='ANZSCO')$results[, 3]} tosee a list of availabledata sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
NA
\link{anzsco}
}
