% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANZSCO.R
\docType{data}
\name{anzsco}
\alias{anzsco}
\title{dataframe with every ANZSCO occupation fully defined}
\format{a \code{data.frame} containing the following fields:
\describe{
\item{Major.Group}{}
\item{Major.Group.Code}{}
\item{Sub.Major.Group.Code}{}
\item{Sub.Major.Group}{}
\item{Minor.Group.Code}{}
\item{Minor.Group}{}
\item{Unit.Group.Code}{}
\item{Unit.Group}{}
\item{Occupation}{}
\item{Occupation.Code}{}
}}
\source{
The data comes from http://www.abs.gov.au/ANZSCO
}
\description{
Detailed description of the data
}
\seealso{
\link{ANZSCO-package}
}
